Balance field -> float

+++ internal/dbprovider/dbprovider.go
@@ -11,26 +11,26 @@ import (
 )
 
 type MessageUser struct {
-	Id      int64  `json:"id"`
-	Name    string `json:"name"`
-	Age     int64  `json:"age"`
-	Email   string `json:"email"`
-	Balance int64  `json:"balance"`
+	Id      int64   `json:"id"`
+	Name    string  `json:"name"`
+	Age     int64   `json:"age"`
+	Email   string  `json:"email"`
+	Balance float64 `json:"balance"`
 }
 
 type PostMessageUser struct {
-	Id            int64  `json:"id"`
-	Sum           int64  `json:"sum"`
-	Action        string `json:"action"`
-	DestinationId int64  `json:"destination"`
+	Id            int64   `json:"id"`
+	Sum           float64 `json:"sum"`
+	Action        string  `json:"action"`
+	DestinationId int64   `json:"destination"`
 }
 
 type Manager interface {
 	AddUser(usr *MessageUser) error
 	GetBalance(id uint64) []byte
-	AddBalance(id uint64, sum uint64) []byte
-	SubBalance(id uint64, sum uint64) []byte
-	SendBalance(id uint64, destId uint64, sum uint64) []byte
+	AddBalance(id uint64, sum float64) []byte
+	SubBalance(id uint64, sum float64) []byte
+	SendBalance(id uint64, destId uint64, sum float64) []byte
 	// FindUser(usr *MessageUser) (*MessageUser, error)
 }
 
@@ -67,7 +67,7 @@ func (mgr *manager) GetBalance(id uint64) []byte {
 	}
 	return js
 }
-func (mgr *manager) AddBalance(id uint64, sum uint64) []byte {
+func (mgr *manager) AddBalance(id uint64, sum float64) []byte {
 	var usr model.User
 	fmt.Println(id)
 	mgr.db.Model(&usr).Where("Id = ?", id).Update("Balance", gorm.Expr("Balance + ?", sum))
@@ -80,11 +80,11 @@ func (mgr *manager) AddBalance(id uint64, sum uint64) []byte {
 	return js
 }
 
-func (mgr *manager) SubBalance(id uint64, sum uint64) []byte {
+func (mgr *manager) SubBalance(id uint64, sum float64) []byte {
 	var usr model.User
 	mgr.db.First(&usr, "Id = ?", id)
 
-	if usr.Balance-int64(sum) > 0 {
+	if usr.Balance-sum > 0 {
 		mgr.db.Model(&usr).Where("Id = ?", id).Update("Balance", gorm.Expr("Balance - ?", sum))
 		js, err := json.Marshal(MessageUser{int64(usr.Id), usr.Name, int64(usr.Age), usr.Email, usr.Balance})
 		if err != nil {
@@ -101,12 +101,12 @@ func (mgr *manager) SubBalance(id uint64, sum uint64) []byte {
 	}
 }
 
-func (mgr *manager) SendBalance(id uint64, destId uint64, sum uint64) []byte {
+func (mgr *manager) SendBalance(id uint64, destId uint64, sum float64) []byte {
 	var usr model.User
 	var usr2 model.User
 	mgr.db.First(&usr, "Id = ?", id)
 
-	if usr.Balance-int64(sum) > 0 {
+	if usr.Balance-sum > 0 {
 		mgr.db.Model(&usr).Where("Id = ?", id).Update("Balance", gorm.Expr("Balance - ?", sum))
 		mgr.db.First(&usr2, "Id = ?", destId)
 		mgr.db.Model(&usr2).Where("Id = ?", destId).Update("Balance", gorm.Expr("Balance + ?", sum))
